+++ xml_generator/Cargo.toml
@@ -5,7 +5,6 @@ edition = "2024"
 
 [dependencies]
 xsd-parser = "1.2.0"
-syn = { version = "2.0.104", features = ["full"] }
 xml-builder = "0.5.4"
 fake = "4.3.0"
